fix: failing tests
