+++ test/credentials/presentations/issuance/ld_vp_v2_test.dart
@@ -15,8 +15,8 @@ void main() async {
 
   final ldV1VC = UniversalParser.parse(VerifiableCredentialDataFixtures
       .credentialWithValidProofDataModelV11JsonEncoded);
-  final ldV2VC = UniversalParser.parse(
-      VerifiableCredentialDataFixtures.credentialWithProofDataModelV20String);
+  final ldV2VC = UniversalParser.parse(VerifiableCredentialDataFixtures
+      .credentialWithValidProofDataModelV20String);
   final sdjwtV2VC =
       UniversalParser.parse(VerifiableCredentialDataFixtures.sdJwtWithValidSig);
